requirements.txt
Django==1.11
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-import-export==1.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
future==0.16.0
drf-extensions==0.4.0
coreapi==2.3.3
pymsql==0.9.2
